BadgesGenerator/requirements.txt
Pillow==8.1.2
qrcode==6.1
